refactor: move streaming logic to dedicated modules for each provider
This change extracts the streaming implementations for Anthropic, Google, and OpenAI into separate files under a new `streaming` directory. The main `llm_requests.py` file now imports these functions and routes requests accordingly. This improves code organization and maintainability by separating provider-specific logic.
